EPMLABSBRN-304 move player logic into service
