+++ frontend/app/components/task-player/sentence/component.js
@@ -77,19 +77,19 @@ export default class SentenceComponent extends Component {
   }
 
   async runNextTaskTimer() {
-    this.args.onRightAnswer();
     await customTimeout(3000);
     if (this.task.isLastTask) {
       this.showExerciseResult();
       await customTimeout(3000);
     }
+    this.args.onRightAnswer();
   }
 
   async handleWrongAnswer() {
     await customTimeout(1000);
     this.task.set('repetitionCount', this.task.repetitionCount + 1);
-    this.audio.player.playAudio();
     this.currentAnswerObject = null;
+    this.args.onWrongAnswer();
   }
 
   async handleCorrectAnswer() {
