+++ frontend/app/components/task-player/sentence/template.hbs
@@ -4,7 +4,6 @@
   {{else}}
     <AudioPlayer
       @audioFileUrl={{this.audioFiles}}
-      @autoplay={{@autoplay}}
       @disabled={{@disableAudioPlayer}}
     />
     <div class="flex mt-4 justify-center">
