+++ frontend/app/components/task-player/single-words/template.hbs
@@ -14,7 +14,6 @@
 {{else}}
   <AudioPlayer
     @audioFileUrl={{@task.audioFileUrl}}
-    @autoplay={{@autoplay}}
     @disabled={{@disableAudioPlayer}}
   />
   <ul class="task-player__options mt-5">
