+++ frontend/app/components/task-player/words-sequences/template.hbs
@@ -3,7 +3,6 @@
   {{#if this.tasksCopy.length}}
     <AudioPlayer
       @audioFileUrl={{this.audioFiles}}
-      @autoplay={{@autoplay}}
       @disabled={{@disableAudioPlayer}}
     />
     <div class="flex mt-4 justify-center">
